+++ code/src/Loading.js
@@ -1,0 +1,32 @@
+import React from "react";
+import loadingimg from "./images/Loading.mp4"
+
+const Loading = () => {
+	return (
+    <div className="pageNotFoundContainer">
+
+
+<div className="imgPageNotFound">
+<div class="center">
+  <div class="wave"></div>
+  <div class="wave"></div>
+  <div class="wave"></div>
+  <div class="wave"></div>
+  <div class="wave"></div>
+  <div class="wave"></div>
+  <div class="wave"></div>
+  <div class="wave"></div>
+  <div class="wave"></div>
+  <div class="wave"></div>
+</div>
+</div>
+
+      <div className="textPageNotFound"> 
+      <h2>Loading....</h2>
+      </div>
+
+    </div>
+  );
+};
+
+export default Loading;
