+++ code/src/Toprated.js
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from "react";
 import { Link } from "react-router-dom";
-import Loading from "components/Loading";
+import Loading from "Loading";
 import { TopRatedUrl } from "Urls";
 
 const Toprated = () => {
